beta-7.2
Had to adjust the way that the util script was looking for and reading data.yaml. Now, alongside training, data.yaml is automatically stored in the model's associated run folder. Now, it can be easily model specific without the user having to input anything, so long as the data.yaml is within this folder.

Signed-off-by: kgoertle <startingupsterling@gmail.com>