beta-7.9
After so much trial and error, all of the pathing logic was placed into paths.py. This now also fully supports reading model-specific data.yaml files and generating a new classes_config.yaml file for user configuration of FOCUS and CONTEXT classes. This enables for "Objects" to be summed and interactions to be more controlled.

Signed-off-by: kgoertle <startingupsterling@gmail.com>