beta-7.9
Updated classes_config.py in order to fully consolidate pathing logic into paths.py. This update continues the progress toward modularity with multiple possible models that a user could want to use with the pipeline. For this, classes_config.yaml now fully self generates if it does not exist and parses the current model's class list into the configs folder.

Signed-off-by: kgoertle <startingupsterling@gmail.com>