beta-7.10
This update includes the ability for the pipeline to use model-specific classes_config.yaml files with safe handling if it isn't setup properly. For clarity on this, if the .yaml file is not configured properly, it will simply reload the config and use for how it is normally created. This allows for far more modularlity in the overall pipeline, more easily allowing users to run multiple models through it.

Signed-off-by: kgoertle <startingupsterling@gmail.com>