beta-8.0
Massive update to the overall design of the downloads. Now, three YOLO versions are available for a user to choose. The pipeline automatically checks the system for the specified model file (.pt or .yaml) and automatically downloads them if they are not present. This helper util also ensures that files are not redownloaded if they already exist.

For reference, this also holds the definitions for the simplified names of the YOLO model files. So, instead of having to write "yolov8n.pt" for -m, a user can just write "-m yolov8" for simplicity.

Signed-off-by: kgoertle <startingupsterling@gmail.com>