beta-8.1
Beta-8.1 updates the way that the val_split.py function works for detecting and automatically splitting label-studio project folders.

Now, the parent directory will contain a folder called "labelstudio-projects". Any folder containing the label-studio export folder structure will automatically be detected, copied, and split into the "data" folder with an accompany metadata.json file. The metadata contains information regarding when the data processing occurred, the original name of the project, and the processed name.

Data folders now follow the same name as the name set for the model itself. So, if a user names their model "lizards", it will name the dataset folder "lizards" as well. This allows for added clarity in model-dataset alignment. 

The validation split also checks labels for whether or not they are for OBB or standard YOLO HBB models, which is also stored in the metadata file. 
The pipeline will check for if a dataset is OBB or HBB and doesn't allow for dataset-model mismatches. So, if a user has an OBB dataset but uses `-m yolo11`, it will fallback to YOLO11-OBB. 

Though, worth noting, that if this mismatch occurs while trying to train another model version, like `-m yolo12` with an OBB dataset, it will always fallback to YOLO11 model weights/architecture. This is just a safety protocol for users so that the safest, most precise model is used as fallback. Regardless, there will ALWAYS be a dataset-model match and training will initiate.

Signed-off-by: kgoertle <startingupsterling@gmail.com>