beta-8.1
In summary, the config now allows for fallback to safe YOLO11 model for incorrect dataset-model handling. As an example, if the user's dataset is oriented for OBB models and they try to run a non-OBB model over it, it will fallback to the OBB version to safely go through training. 
This acts as a safety net for inexperienced users.

`--update` has also been patched to align with the selected model's dataset based on its associated data.yaml path definition. This gives far more modularity and cohesion with the entire update functionality.

This sets the parallel naming structure for data/dataset → runs/model-name → logs/model-name. Now, all folders are model specific. 

Signed-off-by: kgoertle <startingupsterling@gmail.com>