beta-8.1
Massively updated the overall functionality of the validation split functionality.

For clarity, "validation split" refers to how a label-studio project's images and labels folder must be split into a "train" and "val" folder so that training can have image references to validate its precision and accuracy. 

With this patch, users should now place their newly exported label-studio projects into "labelstudio-projects". This helper util will automatically check this folder for new projects and process them into the "data" folder, with the images and labels being split between the necessary "train" and "val" folders.
This comes alongside "data.yaml" being generated with it so that training can use it to define where the dataset is. 

Using `--name` or `-n` to set a model's name is now paralleled in the dataset folder's name. With these changes, now all of the associated model folders are named in parallel. 
As an example, if a user names a new model "geckos", the model folder within "runs" will be called "geckos", the associated model folder within "logs" will be called "geckos", and the dataset within "data" will be called "geckos". The redundancy is integral to allow for users to experiment with multiple models within their directory.

I am assuming that this will be great for not only models using entirely different classes (bird classes vs squirrel classes, as an example), but also for if they're experimenting with multiple model types. They can name a model "squirrels-OBB" to not only label the class contents, but also what type of model it was. If not, the metadata holds what model type was used anyway.



Signed-off-by: kgoertle <startingupsterling@gmail.com>