1.0.2
Initial upload to the new 1.0.0 release. Thank you for everybody who has helped me get to this point, including my lab director, my roommates, my wonderful girlfriend, and my friends. I would not be here if it wasn't for your continued and assuring support! 

Note that the structure of the entire directory is now in parallel with the 'yolo4r' package's directory. 


Signed-off-by: kgoertle <startingupsterling@gmail.com>