1.1.0
Pressing on to major 1.1.0 release, included is a refractor of detect.py and enhanced terminal logging system. 

detect.py was refractored for organization and stability, where in video utility and inference functions were separated into their own unique helper functions. Now, the VideoProcessor receives video encoding and box formatting information from the respective helpers.

The terminal logging system was massively improved in order to support window scaling, class list columns, and reorganizes how the information is section.
Now, there is a dedicated section to the [source] section that allows for classes and their counts to be formatted as a list with dynamic columns for readability. [INFO] prints are now within their own section below the [source] section, with all [SAVE] information being set to its own section, as well. 

This works beautifully with running full-scale official YOLO models so that the terminal doesn't break. This also sets us up to integrate running multiple models themselves in tandem, along with multiple sources at once, as is the current behavior.

Signed-off-by: kgoertle <startingupsterling@gmail.com>