1.1.6
FINALLY updated the entire UI for BOTH the training and detection sides of the entire pipeline. In doing so, now both sides have their terminal printing logic nested inside "console.py". 

This allows for a far, FAR more robust UI layout. This design is now stable and includes chunking sections of the terminal log for readability and organization. The readability is now FAR greater, with each training and detection run clearing and setting specific regions for the aforementioned chunks.

This is now as close as the entire pipeline has ever gotten to being a fully-fledged training-detection pairing pipeline. 

Signed-off-by: kgoertle <startingupsterling@gmail.com>